Fixed a weird bug.
In only some events in some run, due to an empty pointer to a TLorentzVector, the job was crushing.
This happened calling

         if ( track.IsInJet() ) {
            //Info("Process()", "Track is a candidate and within jet: %f <= %f. Saving jet...", track.mMinDeltaRToJet, sMinTrackIsoDeltaR);
            //utils::PrintTLorentzVector((TLorentzVector&) *track->mJet);
            mJetsWithIsoTrack.insert(track.mJet);  // a bug?!?!?!
         }

This caused the tree not to be produced for the entire run. Now I comment this out, since is not used in the analysis.

The scrush huppened like it follows:

Program received signal SIGSEGV, Segmentation fault.
0x005bf156 in TLorentzVector::T (this=0x0) at /afs/rhic.bnl.gov/star/ROOT/5.34.30/.sl64_gcc482/rootdeb/include/TLorentzVector.h:280
280    inline Double_t TLorentzVector::T() const { return fE; }
(gdb) backtrace
    __k=@0xab151d4) at /opt/gcc/4.8.2/include/c++/4.8.2/bits/stl_tree.h:1324
    __v=@0xab151d4) at /opt/gcc/4.8.2/include/c++/4.8.2/bits/stl_tree.h:1377
    at /opt/gcc/4.8.2/include/c++/4.8.2/bits/stl_set.h:463